requirements.txt
"""
Requirements file for the Audio Summarizer project.
Install with: pip install -r requirements.txt
"""

# Core dependencies
openai-whisper>=20230314
llama-cpp-python>=0.2.0
torch>=2.0.0
torchaudio>=2.0.0

# Optional: for CUDA support (uncomment if you have NVIDIA GPU)
# torch>=2.0.0+cu118
# torchaudio>=2.0.0+cu118

# Development dependencies (optional)
# black>=23.0.0
# flake8>=6.0.0
# mypy>=1.0.0
